:sparkles: Renamed UserProfileDetail to UserCardDetail
