+++ pride/babel.config.js
@@ -20,7 +20,8 @@ module.exports = function (api) {
             '@context': './src/context'
           }
         }
-      ]
+      ],
+      'react-native-reanimated/plugin'
     ]
   };
 };
