+++ pride/src/components/UserCardDetail/index.tsx
@@ -1,0 +1,4 @@
+import UserCardDetail from './UserCardDetail';
+
+export * from './UserCardDetail';
+export default UserCardDetail;
