--- pride/src/components/UserProfileDetail/index.tsx
@@ -1,4 +1,0 @@
-import UserProfileDetail from './UserProfileDetail';
-
-export * from './UserProfileDetail';
-export default UserProfileDetail;
